+++ igor-web/src/main/groovy/com/netflix/spinnaker/igor/exceptions/ArtifactNotFoundException.java
@@ -23,8 +23,7 @@ import org.springframework.web.bind.annotation.ResponseStatus;
 
 @ResponseStatus(NOT_FOUND)
 public class ArtifactNotFoundException extends NotFoundException {
-  public ArtifactNotFoundException(
-      String master, String job, Integer buildNumber, String fileName) {
+  public ArtifactNotFoundException(String master, String job, Long buildNumber, String fileName) {
     super(
         String.format(
             "Could not find build artifact matching requested filename '%s' on '%s/%s' build %s",
